Create new Job Portal application with initial entity and services
This commit includes the creation of Job Portal application entities (Job, Applicant, Application) and their corresponding services. It also establishes configurations for Maven, Spring Boot and JPA. Further, a User Details Service for Spring Security and a basic authentication setup have also been added for user management.

This sets up the initial structure of the application, which will be further built on to provide a comprehensive job portal service.